// packages/flows/packages/flow-content/workflow/lib/types/node-type.ts
export enum CustomNodeType {
    KnowledgeBase = 'KnowledgeBase',
    LLM = 'LLMClient',
    IntendClassification = 'IntendClassification',
    DocumentExtractor = 'DocumentExtractor',
    RPC = 'RPC',
    Tool = 'AgentTool',
    Agent = 'Agent',
    MCP = 'MCP',
    HttpRequest = 'HttpRequest',
    TemplateConverter = 'TemplateConvertor',
}